enhance(glean): add http_status metric to pings
currently only either "200" or "404"

merged HydrationData types, and typed pageNotFound property
