+++ client/src/app.tsx
@@ -25,6 +25,7 @@ import { useGleanPage } from "./telemetry/glean-context";
 import { MainContentContainer } from "./ui/atoms/page-content";
 import { Loading } from "./ui/atoms/loading";
 import { Advertising } from "./advertising";
+import { HydrationData } from "../../libs/types/hydration";
 
 const AllFlaws = React.lazy(() => import("./flaws"));
 const Translations = React.lazy(() => import("./translations"));
@@ -108,9 +109,15 @@ function PageOrPageNotFound({ pageNotFound, children }) {
   );
 }
 
-export function App(appProps) {
+export function App(appProps: HydrationData) {
+  const { pathname } = useLocation();
+  const initialPathname = React.useRef(pathname);
+  const pageNotFound =
+    (appProps.pageNotFound || false) && initialPathname.current === pathname;
+
   usePing();
-  useGleanPage();
+  useGleanPage(pageNotFound);
+
   const localeMatch = useMatch("/:locale/*");
 
   useEffect(() => {
@@ -118,18 +125,6 @@ export function App(appProps) {
 
     document.documentElement.setAttribute("lang", locale);
   }, [appProps.locale, localeMatch]);
-
-  const [pageNotFound, setPageNotFound] = React.useState<boolean>(
-    appProps.pageNotFound
-  );
-  const { pathname } = useLocation();
-  const initialPathname = React.useRef(pathname);
-
-  React.useEffect(() => {
-    setPageNotFound(
-      appProps.pageNotFound && initialPathname.current === pathname
-    );
-  }, [appProps.pageNotFound, pathname]);
 
   const isServer = useIsServer();
