+++ client/src/homepage/featured-articles/index.tsx
@@ -1,6 +1,6 @@
 import useSWR from "swr";
 import { CRUD_MODE } from "../../env";
-import { HydrationData } from "../../types/hydration";
+import { HydrationData } from "../../../../libs/types/hydration";
 
 import "./index.scss";
