--- client/src/types/hydration.ts
@@ -1,5 +1,0 @@
-interface HydrationData<Type = any> {
-  hyData: Type;
-}
-
-export type { HydrationData };
